requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
flask==3.1.0; python_version >= '3.9'
gunicorn==23.0.0; python_version >= '3.7'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
python-dotenv==1.1.0; python_version >= '3.9'
pytodotxt==3.0.0; python_version >= '3.7'
werkzeug==3.1.3; python_version >= '3.9'
